+++ tests/templatetags/test_tags.py
@@ -1,5 +1,7 @@
 from django import template
 
+from lazy_tags.decorators import lazy_tag
+
 
 register = template.Library()
 
@@ -7,6 +9,27 @@ register = template.Library()
 @register.simple_tag
 def test():
     return '<p>hello world</p>'
+
+
+@register.simple_tag
+@lazy_tag
+def test_decorator():
+    return 'Success!'
+
+
+@register.simple_tag
+@lazy_tag
+def test_simple_dec_args(arg, kwarg=None):
+    return '{0} {1}'.format(arg, kwarg)
+
+
+@register.inclusion_tag('tests/decorator_tag_with_args.html')
+@lazy_tag
+def test_decorator_with_args(arg, kwarg=None):
+    return {
+        'arg': arg,
+        'kwarg': kwarg
+    }
 
 
 @register.simple_tag